cloud_function/requirements.txt
pysftp~=0.2.9
google-cloud==3.0.0